Moving evaluate function to social history service and updating dockerfile with converter release
