+++ containers/ecr-viewer/src/app/view-data/services/evaluateFhirDataService.tsx
@@ -8,12 +8,10 @@ import {
   Element,
   Encounter,
   Location,
-  Observation,
   Organization,
   Patient,
   Practitioner,
   PractitionerRole,
-  QuestionnaireResponse,
   RelatedPerson,
 } from "fhir/r4";
 import { DateTime } from "luxon";
@@ -77,6 +75,7 @@ import {
   evaluateExposureDetails,
   evaluateTravelHistoryTable,
   returnDisabilityStatusTable,
+  evaluateSocialDeterminantsOfHealth,
 } from "./socialHistoryService";
 import {
   FhirIndex,
@@ -695,141 +694,6 @@ export const evaluateOccupationHistory = (fhirBundle: Bundle) => {
           expanded: false,
           content,
           id: obs.id || `${Math.random()}`,
-          headingLevel: "h5",
-        };
-      })}
-    />
-  );
-};
-
-/**
- * Evaluates Social Determinants of Health (SDOH) from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing SDOH data.
- * @returns An array of evaluated and formatted SDOH questionnaire data.
- */
-export const evaluateSocialDeterminantsOfHealth = (fhirBundle: Bundle) => {
-  const socialFuncObs = evaluateAll(
-    fhirBundle,
-    fhirPathMappings.historyOfSocialFunction,
-  );
-
-  if (socialFuncObs.length === 0) return;
-
-  return (
-    <ExpandCollapseAccordion
-      className="accordion-rr"
-      descriptor="social determinants of health"
-      items={socialFuncObs.map((socialFunc) => {
-        const domainRef = evaluateOne(socialFunc, fhirPathMappings.hasMember);
-
-        const domain = evaluateReference<Observation>(fhirBundle, domainRef);
-
-        const questionnaireResponsesRefs = evaluateAll(
-          domain,
-          fhirPathMappings.observationDerivedFrom,
-        );
-
-        const domainQuestionsAndAnswers = questionnaireResponsesRefs.map(
-          (ref) => {
-            const questionnaireResponse =
-              evaluateReference<QuestionnaireResponse>(fhirBundle, ref);
-
-            const items = evaluateAll(
-              questionnaireResponse,
-              fhirPathMappings.questionnaireItem,
-            );
-
-            const questionsAndAnswers = items.map((item, j) => {
-              const question = item.text;
-
-              const answers = item.answer || [];
-              const answer = answers
-                .map((a) =>
-                  evaluateValue(
-                    a,
-                    fhirPathMappings.valueX,
-                    "QuestionnaireResponse.item.answer",
-                  ),
-                )
-                .join("\n");
-
-              return {
-                Question: {
-                  value: question,
-                },
-                Answer: {
-                  value: answer,
-                },
-              } as HtmlTableJsonRow;
-            });
-
-            return questionsAndAnswers;
-          },
-        );
-
-        const content = [
-          <JsonTable
-            key={`${domain?.id}-questions-and-answers`}
-            jsonTableData={{ tables: domainQuestionsAndAnswers }}
-            className="caption-data-title margin-y-0"
-            outerBorder={false}
-            columnStyles={{
-              0: { width: "200px", minWidth: "100px" }, // First column (Question)
-              1: { width: "80px", minWidth: "80px" }, // Second column (Answer)
-            }}
-          />,
-        ];
-
-        const h6ClassName =
-          "bg-gray-5 margin-x-neg-205 padding-y-2 padding-x-205";
-        content.push(
-          <h6
-            key={`${domain?.id}-finding-title`}
-            // inline styling to overwrite usa-prose nested style
-            style={{
-              marginTop: "-1rem",
-              fontWeight: "bold",
-              borderBottom: "1px solid black",
-            }}
-            className={h6ClassName}
-          >
-            Available Social Determinants of Health Information
-          </h6>,
-        );
-
-        const findings = domain?.interpretation?.map((item, i) => {
-          const riskValue = item.text
-            ? item.text
-            : evaluateValue(item, fhirPathMappings.codingDisplay);
-
-          return {
-            title: "Finding",
-            value: riskValue,
-          } as DisplayDataProps;
-        });
-
-        content.push(
-          <DataDisplayList
-            key={`${domain?.id}-findings`}
-            items={findings ?? []}
-          />,
-        );
-
-        const domainTitle = evaluateValue(domain, fhirPathMappings.code);
-
-        return {
-          title: (
-            <div className="display-flex flex-row flex-no-wrap flex-justify">
-              <span>{domainTitle}</span>
-
-              <span className="font-size-xs text-base">
-                {evaluateValue(domain, fhirPathMappings.effectiveX)}
-              </span>
-            </div>
-          ),
-          expanded: false,
-          content,
-          id: socialFunc.id ?? "ID goes here",
           headingLevel: "h5",
         };
       })}
